Added job template, tested 1 file, submitted all Ele data jobs
Commented out lumi and MET filters in NtuplerAnalyzer.cc
The lumi is done in python (not clear how it reports how much lumi is processed).
MET filters -- tried python recipe, it crashed not finding some EcalDeadCells (yep, there is such a filter)
-- probably outdated and some names/types changed.

Submitted all Ele data anyway for the test of volume (now record on res-lep/di-lep and 1 b-tag).
And to test jobs running on network.

	modified:   plugins/NtuplerAnalyzer.cc
	new file:   python/LSF_template_ConfFile_cfg.py