Clean up, phase 2
Now the jets and taus should be cleanly done.
Couple fixes are made along the way.
Jets and taus are pt sorted -- not sure if it is really needed, it should be done in ntupler.
The jets are cross-cleaned from the first tau in corresponding selection.

Selection defines cuts and IDs and everything.

	modified:   support_channel_distrs.py