Fixed tau-cleaning (most of selections failed over it), changed the jet cor propagation to met, increased lj cut 50->80
And ran test jobs on it.

And added the tau-matched jets into the calculation of lj variable.

The saved jets were not corrected with JER,
they = slimmedJets.
And I checked UncocFactor for multiplication -- it does equalize to UncorJet.
Now nominal factor is jet * uncor * jer * jes = jet * en_factor
and this is propagated to met accordingly.

Additional control distrs:
for sum(particles+met),
for Mt at small amount of jets and at the bin of 0L 2M b-tagged jets (the most deviation in tau sel for N jets histograms)
etc.

Also added options to run of normalizations in parallel across files.

	modified:   jobsums/mc_norm/normalize.py
	modified:   jobsums/norm_to_mu/normalize.py
	modified:   plugins/NtuplerAnalyzer.cc
	modified:   support_channel_distrs.py