Checked the latest selection agains old results
The selection for dy-ltau seems to underperform.
However, the comparison of cumulative fractions to old selections,
in quick tests of v40u2 and the old v14raw2, show the same fraction of wjets in the tail of sv sign.
Probably, the jets in W production do come from c quarks often enough, which is bad news.
Bad at the same time now we have a large selection of true taus and W fakes to finish the study of dalitz parameters.

	new file:   studies/dy_mutau_tests/dy_mutau_tests_quick_v40u2_inputs.png
	new file:   studies/dy_mutau_tests/dy_mutau_tests_quick_v40u2_inputs_Mt40.png
	new file:   studies/dy_mutau_tests/foobar_v40u2_inputs_Mt40_wjets_tail_cumulative-fractions.png
	new file:   studies/dy_mutau_tests/foobar_v40u2_inputs_wjets_tail_cumulative-fractions.png
	new file:   studies/dy_mutau_tests/old_results_v14raw2_mc_wFSRISR_mu_ctr_mu_dy_tt_cumulative-fractions.png
	new file:   studies/dy_mutau_tests/old_results_v14raw2_mc_wFSRISR_mu_ctr_mu_dy_tt_DY-tautau-yield-cumulative.png