And finish of main paradox in pT trends
at least in dataH there is no significant deviation from v12 now
-- the rest is the damned difference in full and GH datasets?

The step in e pT-s remains, but it is probably due to one of SFs
(one 2D histo, when there are better 1D ones?)
It might be the cause for the slight trends in non-tt MC processes
It could also be the cause of the deviation in high etas.
And the jums in etas miht be due to not so precise 2D SF histo,
when 1D are asymmetric.

	modified:   last_procs
	modified:   stage2.py