Thorough investigation of asymmetry in the fit of both channels
aparently the issue is due to TOPPT and other systematics correlating with it:
fit in mu does not prefer TOPPT, but el does,
then the fit in both does prefer TOPPT again,
other systematics adjust remaining differences.

Frag is found to be too big, which also showed up previously in sys yield normalizations.
Aparently, it has changed in the ntuples, even though there is no difference found neither in Ntupler nor in Bfragmentation.
(Although there is some backup BfragmentationWeight.cc file with additional conditions. It's not clear if the conditions affect something significantly.)
The effect of new Frag weights is that the variation is about 1/4 larger than what it was before (2% -> 2.5% etc).
Reducing Frag does help with the asymmetry in the fit, but does not remove it completely.

	modified:   daily_scripts
	modified:   proc/jobsums/Makefile
	modified:   syst_rates.py